New files for clarity of what's where
